call PlanResourceChange for destroy operations
Previously Terraform would create destroy plans for resources without
any input from the provider, because a destroy change always results in
a null value. While the change alone is sufficient for Terraform to
generate the plan containing destroy operations, it does not allow
providers any way to check the validity of the change.

Here we add in a call to PlanResourceChange whenplanning resource
destroy actions, sending null proposed and config values to indicate
that this is a destroy operation. This gives the option for providers to
return any diagnostics about the operation, particularly when it can
detect if the destroy is going to fail during apply.